chore(relay): address PR #728 review feedback
Three hygiene fixes from founder review, no logic changes:

- Remove the repeated ADR-037 P2 explanatory comment block from the 5
  memory-read call sites (context.rs, memory/{list,search,show,timeline}.rs).
  The canonical explanation already lives on outbox::poll_and_apply's doc
  comment; only the call sites' local duplicate copies are removed.
- Replace the "founder review (PR #728)" attribution comment in
  handlers.rs's dedupe-skip test section with the plain suggested text.
- Split relay.rs's #[cfg(test)] module into its own file, following the
  dedupe/entity_id_migration precedent: relay.rs -> relay/{mod.rs,tests.rs}.
  Pure move; same 10 tests before and after.
